Decompose PowerMap.jsx into a powermap/ module directory (behavior parity)
Split the 540-line deck.gl map component into focused modules under
frontend/src/components/powermap/: palettes, constants, scales, a minimal
fill registry (structural scaffolding — color fns move in later PRs), four
layer factories, useMapData, tooltip builder, Legend rows, Scrubber, and an
index.jsx container. Every constant and load-bearing comment moved with its
code; all updateTriggers carried over unchanged.

Deliberate changes beyond the move:
- useMapData switches the three API GETs to useFetchWithError (SWR +
  URL-keyed in-flight dedupe: /power/overview is now physically shared with
  PowerOverviewMatrix/NarrativeHero, /power/borders with BordersPanel). No
  transform, matching the other consumers of the same URLs, so the shared
  cache keeps one payload shape; no pollMs (the map stays one-shot at
  mount). The static geojson stays a plain fetch.
- State renames: metric -> fill (driven by the FILLS registry), showArcs ->
  overlays.flows.
- New safe-optional props for PR 8: onZoneSelect (zone click on polygons/
  points; context countries never fire), selectedZone (accent outline on
  ALL matching features — DE_LU/FR span two each), tall (viewport-height
  canvas; default keeps the fixed 460px).

App.jsx only re-points the lazy import; the old monolith is removed.
